requirements.txt
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
aiohttp==3.9.1
python-decouple==3.8
tavily-python
